:elf: Use elf store for theming
