+++ components/styles/ThemeDropdown.tsx
@@ -1,8 +1,9 @@
 import Image from 'next/image';
 import { Fragment, FunctionComponent } from 'react';
 import { Listbox, Transition } from '@headlessui/react';
-import useThemeProvider from '../../shared/providers/theme.provider';
 import { Theme } from '../../shared/models/theme.model';
+import { useObservable } from '@ngneat/react-rxjs';
+import { themeStore } from '../../shared/stores/theme/theme.store';
 
 interface Props {
   selectedTheme: Theme;
@@ -11,7 +12,7 @@ interface Props {
 }
 
 const ThemeDropdown: FunctionComponent<Props> = ({ selectedTheme, setTheme, themes }) => {
-  const { theme } = useThemeProvider();
+  const [theme] = useObservable(themeStore);
 
   return (
     <Listbox value={selectedTheme} onChange={setTheme}>
