+++ pages/login.tsx
@@ -1,4 +1,5 @@
 import type { NextPage } from 'next';
+import { useObservable } from '@ngneat/react-rxjs';
 import CustomHeader from '../components/core/CustomHeader';
 import LoginNav from '../components/navs/LoginNav';
 import CheckboxInput from '../components/styles/CheckboxInput';
@@ -6,10 +7,10 @@ import NavbarLink from '../components/styles/NavbarLink';
 import PageHeader from '../components/styles/PageHeader';
 import TextInput from '../components/styles/TextInput';
 import LoginHandler from '../shared/handlers/login.handler';
-import useThemeProvider from '../shared/providers/theme.provider';
+import { themeStore } from '../shared/stores/theme/theme.store';
 
 const Login: NextPage = () => {
-  const { theme } = useThemeProvider();
+  const [theme] = useObservable(themeStore);
 
   const {
     username,
