--- shared/providers/theme.provider.ts
@@ -1,19 +1,0 @@
-import { useEffect, useState } from 'react';
-import { Theme } from '../models/theme.model';
-import themeService from '../services/theme.service';
-
-const useThemeProvider = () => {
-  const [theme, setTheme] = useState<Theme>();
-
-  useEffect(() => {
-    const themeSub = themeService.currentTheme$.subscribe(setTheme);
-
-    return (() => {
-      themeSub.unsubscribe();
-    });
-  }, []);
-
-  return { theme };
-};
-
-export default useThemeProvider;
