+++ shared/services/app.service.ts
@@ -5,11 +5,11 @@ import { pairwise } from 'rxjs';
 
 import { itemsService } from '../stores/items/items.service';
 import { userService } from '../stores/user/user.service';
-import themeService from './theme.service';
 import { inventoryService } from '../stores/inventory/inventory.service';
 import { serverNotificationService } from '../stores/serverNotification/server-notification.service';
 import { User } from '../stores/user/user.model';
 import { userStore } from '../stores/user/user.store';
+import { themeService } from '../stores/theme/theme.service';
 
 class AppService {
   public init(): void {
